Clean up simple_endpoint deserialization
