+++ Makefile
@@ -11,9 +11,10 @@ test:
 		testify -v moe.tests
 
 style-test:
-		pip install flake8 flake8-import-order pep8-naming flake8-docstrings
-		flake8 --ignore=E501,E126,E123,I101,I100,,N806 moe
+		pip install flake8 flake8-import-order pep8-naming flake8-docstrings pyflakes
+		flake8 --ignore=E501,E126,E123,I101,I100,N806 moe
 		pep257 moe
+		pyflakes moe
 
 docs:
 		python docs/cpp_rst_maker.py
