+++ moe/views/gp_next_points_pretty_view.py
@@ -5,156 +5,14 @@ Include:
     1. Request and response schemas
     2. Class that extends GpPrettyView for next_points optimizers
 """
-import colander
-
 import numpy
 
-from moe.optimal_learning.python.constant import DEFAULT_EXPECTED_IMPROVEMENT_MC_ITERATIONS, DEFAULT_MAX_NUM_THREADS
 import moe.optimal_learning.python.cpp_wrappers.expected_improvement
 from moe.optimal_learning.python.cpp_wrappers.expected_improvement import ExpectedImprovement
 from moe.views.gp_pretty_view import GpPrettyView
 from moe.views.optimizable_gp_pretty_view import OptimizableGpPrettyView
-from moe.views.schemas import GpHistoricalInfo, ListOfPointsInDomain, CovarianceInfo, BoundedDomainInfo, OptimizationInfo
+from moe.views.schemas import GpNextPointsRequest, GpNextPointsResponse
 from moe.views.utils import _make_gp_from_params, _make_domain_from_params, _make_optimization_parameters_from_params
-
-
-class GpNextPointsRequest(colander.MappingSchema):
-
-    """A ``gp_next_points_*`` request colander schema.
-
-    **Required fields**
-
-        :gp_historical_info: a :class:`moe.views.schemas.GpHistoricalInfo` dict of historical data
-        :domain_info: a :class:`moe.views.schemas.BoundedDomainInfo` dict of domain information
-
-    **Optional fields**
-
-        :num_to_sample: number of next points to generate (default: 1)
-        :mc_iterations: number of Monte Carlo (MC) iterations to perform in numerical integration to calculate EI
-        :max_num_threads: maximum number of threads to use in computation (default: 1)
-        :covariance_info: a :class:`moe.views.schemas.CovarianceInfo` dict of covariance information
-        :optimization_info: a :class:`moe.views.schemas.OptimizationInfo` dict of optimization information
-        :points_being_sampled: list of points in domain being sampled in concurrent experiments (default: [])
-
-    **Example Minimal Request**
-
-    .. sourcecode:: http
-
-        Content-Type: text/javascript
-
-        {
-            "num_to_sample": 1,
-            "gp_historical_info": {
-                "points_sampled": [
-                        {"value_var": 0.01, "value": 0.1, "point": [0.0]},
-                        {"value_var": 0.01, "value": 0.2, "point": [1.0]}
-                    ],
-                },
-            "domain_info": {
-                "dim": 1,
-                "domain_bounds": [
-                    {"min": 0.0, "max": 1.0},
-                    ],
-                },
-        }
-
-    **Example Full Request**
-
-    .. sourcecode:: http
-
-        Content-Type: text/javascript
-
-        {
-            "num_to_sample": 1,
-            "points_being_sampled": [[0.2], [0.7]],
-            "mc_iterations": 10000,
-            "max_num_threads": 1,
-            "gp_historical_info": {
-                "points_sampled": [
-                        {"value_var": 0.01, "value": 0.1, "point": [0.0]},
-                        {"value_var": 0.01, "value": 0.2, "point": [1.0]}
-                    ],
-                },
-            "domain_info": {
-                "domain_type": "tensor_product"
-                "dim": 1,
-                "domain_bounds": [
-                    {"min": 0.0, "max": 1.0},
-                    ],
-                },
-            "covariance_info": {
-                "covariance_type": "square_exponential",
-                "hyperparameters": [1.0, 1.0],
-                },
-            "optimization_info": {
-                "optimization_type": "gradient_descent_optimizer",
-                "num_multistarts": 200,
-                "num_random_samples": 4000,
-                "optimization_parameters": {
-                    "gamma": 0.5,
-                    ...
-                    },
-                },
-        }
-
-    """
-
-    num_to_sample = colander.SchemaNode(
-            colander.Int(),
-            missing=1,
-            validator=colander.Range(min=1),
-            )
-    mc_iterations = colander.SchemaNode(
-            colander.Int(),
-            validator=colander.Range(min=1),
-            missing=DEFAULT_EXPECTED_IMPROVEMENT_MC_ITERATIONS,
-            )
-    max_num_threads = colander.SchemaNode(
-            colander.Int(),
-            validator=colander.Range(min=1),
-            missing=DEFAULT_MAX_NUM_THREADS,
-            )
-    gp_historical_info = GpHistoricalInfo()
-    domain_info = BoundedDomainInfo()
-    covariance_info = CovarianceInfo(
-            missing=CovarianceInfo().deserialize({}),
-            )
-    optimization_info = OptimizationInfo(
-            missing=OptimizationInfo().deserialize({}),
-            )
-    points_being_sampled = ListOfPointsInDomain(
-            missing=[],
-            )
-
-
-class GpNextPointsResponse(colander.MappingSchema):
-
-    """A ``gp_next_points_*`` response colander schema.
-
-    **Output fields**
-
-        :endpoint: the endpoint that was called
-        :points_to_sample: list of points in the domain to sample next (:class:`moe.views.schemas.ListOfPointsInDomain`)
-        :expected_improvement: list of EI of points in points_to_sample (:class:`moe.views.schemas.ListOfExpectedImprovements`)
-
-    **Example Response**
-
-    .. sourcecode:: http
-
-        {
-            "endpoint":"gp_ei",
-            "points_to_sample": [["0.478332304526"]],
-            "expected_improvement": "0.443478498868",
-        }
-
-    """
-
-    endpoint = colander.SchemaNode(colander.String())
-    points_to_sample = ListOfPointsInDomain()
-    expected_improvement = colander.SchemaNode(
-            colander.Float(),
-            validator=colander.Range(min=0.0),
-            )
 
 
 class GpNextPointsPrettyView(OptimizableGpPrettyView):
