+++ moe/views/rest/gp_mean_var.py
@@ -5,264 +5,14 @@ Includes:
     1. request and response schemas
     2. pretty and backend views
 """
-import colander
-
 import numpy
 
 from pyramid.view import view_config
 
 from moe.views.constant import GP_MEAN_ROUTE_NAME, GP_MEAN_PRETTY_ROUTE_NAME, GP_VAR_ROUTE_NAME, GP_VAR_PRETTY_ROUTE_NAME, GP_VAR_DIAG_ROUTE_NAME, GP_VAR_DIAG_PRETTY_ROUTE_NAME, GP_MEAN_VAR_ROUTE_NAME, GP_MEAN_VAR_PRETTY_ROUTE_NAME, GP_MEAN_VAR_DIAG_ROUTE_NAME, GP_MEAN_VAR_DIAG_PRETTY_ROUTE_NAME
 from moe.views.gp_pretty_view import GpPrettyView, PRETTY_RENDERER
-from moe.views.schemas import ListOfFloats, MatrixOfFloats, GpHistoricalInfo, ListOfPointsInDomain, CovarianceInfo, DomainInfo
+from moe.views.schemas import GpMeanVarRequest, GpMeanVarResponse, GpMeanVarDiagResponse, GpMeanResponse, GpVarResponse, GpVarDiagResponse
 from moe.views.utils import _make_gp_from_params
-
-
-class GpMeanVarRequest(colander.MappingSchema):
-
-    """A gp_mean_var request colander schema.
-
-    **Required fields**
-
-        :points_to_sample: list of points in domain to calculate the Gaussian Process (GP) mean and covariance at (:class:`moe.views.schemas.ListOfPointsInDomain`)
-        :gp_historical_info: a :class:`moe.views.schemas.GpHistoricalInfo` object of historical data
-
-    **Optional fields**
-
-        :covariance_info: a :class:`moe.views.schemas.CovarianceInfo` dict of covariance information
-
-    **Example Minimal Request**
-
-    .. sourcecode:: http
-
-        Content-Type: text/javascript
-
-        {
-            "points_to_sample": [[0.1], [0.5], [0.9]],
-            "gp_historical_info": {
-                "points_sampled": [
-                        {"value_var": 0.01, "value": 0.1, "point": [0.0]},
-                        {"value_var": 0.01, "value": 0.2, "point": [1.0]}
-                    ],
-                },
-            "domain_info": {
-                "dim": 1,
-                },
-        }
-
-    **Example Full Request**
-
-    .. sourcecode:: http
-
-        Content-Type: text/javascript
-
-        {
-            "points_to_sample": [[0.1], [0.5], [0.9]],
-            "gp_historical_info": {
-                "points_sampled": [
-                        {"value_var": 0.01, "value": 0.1, "point": [0.0]},
-                        {"value_var": 0.01, "value": 0.2, "point": [1.0]}
-                    ],
-                },
-            "domain_info": {
-                "domain_type": "tensor_product"
-                "dim": 1,
-                },
-            "covariance_info": {
-                "covariance_type": "square_exponential",
-                "hyperparameters": [1.0, 1.0],
-                },
-        }
-
-    """
-
-    points_to_sample = ListOfPointsInDomain()
-    gp_historical_info = GpHistoricalInfo()
-    domain_info = DomainInfo()
-    covariance_info = CovarianceInfo(
-            missing=CovarianceInfo().deserialize({}),
-            )
-
-
-class GpEndpointResponse(colander.MappingSchema):
-
-    """A base schema for the endpoint name.
-
-    **Output fields**
-
-        :endpoint: the endpoint that was called
-
-    **Example Response**
-
-    .. sourcecode:: http
-
-        {
-            "endpoint":"gp_mean_var",
-        }
-
-    """
-
-    endpoint = colander.SchemaNode(colander.String())
-
-
-class GpMeanMixinResponse(colander.MappingSchema):
-
-    """A mixin response colander schema for the mean of a gaussian process.
-
-    **Output fields**
-
-        :mean: list of the means of the GP at ``points_to_sample`` (:class:`moe.views.schemas.ListOfFloats`)
-
-    **Example Response**
-
-    .. sourcecode:: http
-
-        {
-            "mean": ["0.0873832198661","0.0130505261903","0.174755506336"],
-        }
-
-    """
-
-    mean = ListOfFloats()
-
-
-class GpVarMixinResponse(colander.MappingSchema):
-
-    """A mixin response colander schema for the [co]variance of a gaussian process.
-
-    **Output fields**
-
-        :variance: matrix of covariance of the GP at ``points_to_sample`` (:class:`moe.views.schemas.MatrixOfFloats`)
-
-    **Example Response**
-
-    .. sourcecode:: http
-
-        {
-            "var": [
-                    ["0.228910114429","0.0969433771923","0.000268292907969"],
-                    ["0.0969433771923","0.996177332647","0.0969433771923"],
-                    ["0.000268292907969","0.0969433771923","0.228910114429"]
-                ],
-        }
-
-    """
-
-    var = MatrixOfFloats()
-
-
-class GpVarDiagMixinResponse(colander.MappingSchema):
-
-    """A mixin response colander schema for the variance of a gaussian process.
-
-    **Output fields**
-
-        :variance: list of variances of the GP at ``points_to_sample``; i.e., diagonal of the ``variance`` response from gp_mean_var (:class:`moe.views.schemas.ListOfFloats`)
-
-    **Example Response**
-
-    .. sourcecode:: http
-
-        {
-            "var": ["0.228910114429","0.996177332647","0.228910114429"],
-        }
-
-    """
-
-    var = ListOfFloats()
-
-
-class GpMeanResponse(GpEndpointResponse, GpMeanMixinResponse):
-
-    """A gp_mean response colander schema.
-
-    **Output fields**
-
-        :endpoint: the endpoint that was called
-        :mean: list of the means of the GP at ``points_to_sample`` (:class:`moe.views.schemas.ListOfFloats`)
-
-    **Example Response**
-
-    See composing members' docstrings.
-
-    """
-
-    pass
-
-
-class GpVarResponse(GpEndpointResponse, GpVarMixinResponse):
-
-    """A gp_var response colander schema.
-
-    **Output fields**
-
-        :endpoint: the endpoint that was called
-        :variance: matrix of covariance of the GP at ``points_to_sample`` (:class:`moe.views.schemas.MatrixOfFloats`)
-
-    **Example Response**
-
-    See composing members' docstrings.
-
-    """
-
-    pass
-
-
-class GpVarDiagResponse(GpEndpointResponse, GpVarDiagMixinResponse):
-
-    """A gp_var_diag response colander schema.
-
-    **Output fields**
-
-        :endpoint: the endpoint that was called
-        :variance: list of variances of the GP at ``points_to_sample``; i.e., diagonal of the ``variance`` response from gp_mean_var (:class:`moe.views.schemas.ListOfFloats`)
-
-    **Example Response**
-
-    See composing members' docstrings.
-
-    """
-
-    pass
-
-
-class GpMeanVarResponse(GpMeanResponse, GpVarMixinResponse):
-
-    """A gp_mean_var response colander schema.
-
-    **Output fields**
-
-        :endpoint: the endpoint that was called
-        :mean: list of the means of the GP at ``points_to_sample`` (:class:`moe.views.schemas.ListOfFloats`)
-        :variance: matrix of covariance of the GP at ``points_to_sample`` (:class:`moe.views.schemas.MatrixOfFloats`)
-
-    **Example Response**
-
-    See composing members' docstrings.
-
-    """
-
-    pass
-
-
-class GpMeanVarDiagResponse(GpMeanResponse, GpVarDiagMixinResponse):
-
-    """A gp_mean_var_diag response colander schema.
-
-    **Output fields**
-
-        :endpoint: the endpoint that was called
-        :mean: list of the means of the GP at ``points_to_sample`` (:class:`moe.views.schemas.ListOfFloats`)
-        :variance: list of variances of the GP at ``points_to_sample``; i.e., diagonal of the ``variance`` response from gp_mean_var (:class:`moe.views.schemas.ListOfFloats`)
-
-    **Example Response**
-
-    .. sourcecode:: http
-
-    See composing members' docstrings.
-
-    """
-
-    pass
 
 
 class GpMeanVarBaseView(GpPrettyView):
